Refactor directory/cache/deque/index methods to FanoutMixin
